+++ common/src/main/java/dev/qixils/crowdcontrol/common/AbstractPlugin.java
@@ -2,6 +2,7 @@ package dev.qixils.crowdcontrol.common;
 
 import dev.qixils.crowdcontrol.CrowdControl;
 import dev.qixils.crowdcontrol.common.command.Command;
+import dev.qixils.crowdcontrol.common.util.SemVer;
 import dev.qixils.crowdcontrol.socket.SocketManager;
 import lombok.Getter;
 import lombok.Setter;
@@ -52,6 +53,7 @@ public abstract class AbstractPlugin<P, S> implements Plugin<P, S> {
 	protected final ScheduledExecutorService scheduledExecutor = Executors.newSingleThreadScheduledExecutor();
 	@Getter @NotNull
 	protected final Map<String, List<SocketManager>> sentEvents = new HashMap<>();
+	protected final Map<UUID, SemVer> clientVersions = new HashMap<>();
 
 	protected AbstractPlugin(@NotNull Class<P> playerClass, @NotNull Class<S> commandSenderClass) {
 		this.playerClass = playerClass;
@@ -81,9 +83,25 @@ public abstract class AbstractPlugin<P, S> implements Plugin<P, S> {
 			throw new IllegalStateException("CrowdControl is not initialized");
 		try {
 			crowdControl.registerHandler(name, command::executeAndNotify);
-			getSLF4JLogger().debug("Registered CC command '" + name + "'");
+			getSLF4JLogger().debug("Registered CC command '{}'", name);
 		} catch (IllegalArgumentException e) {
-			getSLF4JLogger().warn("Failed to register command: " + name, e);
+			getSLF4JLogger().warn("Failed to register command: {}", name, e);
 		}
+	}
+
+	@Override
+	public @NotNull Optional<SemVer> getModVersion(@NotNull P player) {
+		return Optional.ofNullable(clientVersions.get(playerMapper().getUniqueId(player)));
+	}
+
+	@Override
+	public int getModdedPlayerCount() {
+		return clientVersions.size();
+	}
+
+	@Override
+	public void onPlayerLeave(P player) {
+		clientVersions.remove(playerMapper().getUniqueId(player));
+		Plugin.super.onPlayerLeave(player);
 	}
 }