+++ fabric/src/main/java/dev/qixils/crowdcontrol/plugin/fabric/CommandRegister.java
@@ -3,8 +3,8 @@ package dev.qixils.crowdcontrol.plugin.fabric;
 import dev.qixils.crowdcontrol.common.command.AbstractCommandRegister;
 import dev.qixils.crowdcontrol.common.command.Command;
 import dev.qixils.crowdcontrol.common.command.CommandConstants;
+import dev.qixils.crowdcontrol.common.command.impl.Shader;
 import dev.qixils.crowdcontrol.common.util.MappedKeyedTag;
-import dev.qixils.crowdcontrol.common.util.SemVer;
 import dev.qixils.crowdcontrol.plugin.fabric.commands.*;
 import dev.qixils.crowdcontrol.plugin.fabric.commands.executeorperish.DoOrDieCommand;
 import dev.qixils.crowdcontrol.plugin.fabric.utils.TypedTag;
@@ -103,19 +103,6 @@ public class CommandRegister extends AbstractCommandRegister<ServerPlayer, Fabri
 			() -> GravityCommand.maximum(plugin),
 			() -> new BiomeCommand(plugin),
 			() -> new StructureCommand(plugin),
-			() -> new ShaderCommand(plugin, "bumpy", new SemVer(3, 3, 0)),
-			() -> new ShaderCommand(plugin, "green", new SemVer(3, 3, 0)),
-			() -> new ShaderCommand(plugin, "ntsc", new SemVer(3, 3, 0)),
-			() -> new ShaderCommand(plugin, "desaturate", new SemVer(3, 3, 0)),
-			() -> new ShaderCommand(plugin, "flip", new SemVer(3, 3, 0)),
-			() -> new ShaderCommand(plugin, "invert", new SemVer(3, 3, 0)),
-			() -> new ShaderCommand(plugin, "blobs2", new SemVer(3, 3, 0)),
-			() -> new ShaderCommand(plugin, "pencil", new SemVer(3, 3, 0)),
-			() -> new ShaderCommand(plugin, "sobel", new SemVer(3, 3, 0)),
-			() -> new ShaderCommand(plugin, "cc_wobble", new SemVer(3, 3, 0)),
-			() -> new ShaderCommand(plugin, "bits", new SemVer(3, 3, 0)),
-			() -> new ShaderCommand(plugin, "spider", new SemVer(3, 3, 0)),
-			() -> new ShaderCommand(plugin, "phosphor", new SemVer(3, 3, 0)),
 			() -> new DeleteRandomItemCommand(plugin),
 			() -> new UniteCommand(plugin),
 			() -> TickRateCommand.doubleRate(plugin),
@@ -176,6 +163,10 @@ public class CommandRegister extends AbstractCommandRegister<ServerPlayer, Fabri
 		for (GameType gameType : GameType.values()) {
 			if (gameType == GameType.SURVIVAL) continue;
 			initTo(commands, () -> new GameModeCommand(plugin, gameType, gameType == GameType.SPECTATOR ? 8L : 15L));
+		}
+
+		for (Shader shader : Shader.values()) {
+			initTo(commands, () -> new ShaderCommand(plugin, shader));
 		}
 	}