+++ paper/src/main/java/dev/qixils/crowdcontrol/plugin/paper/commands/EntityCommand.java
@@ -9,7 +9,7 @@ import net.minecraft.world.flag.FeatureFlagSet;
 import org.bukkit.Bukkit;
 import org.bukkit.Difficulty;
 import org.bukkit.World;
-import org.bukkit.craftbukkit.util.CraftMagicNumbers;
+import org.bukkit.craftbukkit.entity.CraftEntityType;
 import org.bukkit.entity.Enemy;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.EntityType;
@@ -28,7 +28,8 @@ public interface EntityCommand extends FeatureElementCommand {
 
 	@Override
 	default @NotNull FeatureFlagSet requiredFeatures() {
-		return CraftMagicNumbers.getEntityTypes(getEntityType()).requiredFeatures();
+		// TODO: API for this is finally available in 1.21 but it's weird
+		return CraftEntityType.bukkitToMinecraft(getEntityType()).requiredFeatures();
 	}
 
 	default boolean isMonster() {