+++ sponge-8/src/main/java/dev/qixils/crowdcontrol/plugin/sponge8/Command.java
@@ -1,17 +1,12 @@
 package dev.qixils.crowdcontrol.plugin.sponge8;
 
 import dev.qixils.crowdcontrol.common.util.RandomUtil;
-import dev.qixils.crowdcontrol.socket.Request;
 import lombok.Getter;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 import org.spongepowered.api.entity.living.player.server.ServerPlayer;
 
-import java.util.List;
 import java.util.Objects;
 import java.util.Random;
-
-import static dev.qixils.crowdcontrol.exceptions.ExceptionUtil.validateNotNullElseGet;
 
 public abstract class Command implements dev.qixils.crowdcontrol.common.command.Command<ServerPlayer> {
 	protected static final Random random = RandomUtil.RNG;
@@ -20,17 +15,5 @@ public abstract class Command implements dev.qixils.crowdcontrol.common.command.
 
 	protected Command(@NotNull SpongeCrowdControlPlugin plugin) {
 		this.plugin = Objects.requireNonNull(plugin, "plugin");
-	}
-
-	@Override
-	public boolean isClientAvailable(@Nullable List<ServerPlayer> possiblePlayers, @NotNull Request request) {
-		if (!plugin.getGame().isClientAvailable())
-			return false;
-		final List<ServerPlayer> players = validateNotNullElseGet(possiblePlayers, () -> plugin.getPlayers(request));
-		if (players.size() != 1)
-			return false;
-		return plugin.getGame().client().player()
-				.map(player -> player.uniqueId().equals(players.get(0).uniqueId()))
-				.orElse(false);
 	}
 }